feat: APIModel 생성
-PostItem: POST할 때 사용될 모델
-PatchItem: PATCH할 때 사용될 모델
-DeleteItem: DELETE할 때 사용될 모델
