+++ OpenMarket/OpenMarket/Model/APIModel/DeleteItem.swift
@@ -1,0 +1,12 @@
+//
+//  DeleteItem.swift
+//  OpenMarket
+//
+//  Created by 오승기 on 2021/08/13.
+//
+
+import Foundation
+
+struct DeleteItem: Codable {
+  var password: String
+}
